requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
dacite==1.8.1
dataclasses-json==0.6.3
Django==4.2.8
django-redis==5.4.0
djangorestframework==3.14.0
drf-spectacular==0.27.0
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
marshmallow==3.20.1
mypy-extensions==1.0.0
packaging==23.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.0
rpds-py==0.13.2
sqlparse==0.4.4
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
